requirements.txt
langchain
langchain_groq
pandas
numpy
pulp
streamlit
redis
python-dotenv
requests
